refactor: move object utilities to dedicated `utils/object` module
